Ensure that messaging is imported for a few failing tests
These tests usually pass when run in bulk because *preceding* tests perform
the necessary import, but when run on their own they reliably fail. The extra
import is necessary so that when you access a message read/write port, SWIG
understands how to wrap it for appropriate use (i.e. that subscribeTo exists
and whatnot). Without the import, such objects are left as bare PySwigObjects.