Remove misleading initialization of xBar whose value is never actually used
The value of xBar is only used by SRukfInterface and its clients, and is always
reset to a new value whenever timeUpdate is called. Since timeUpdate always
occurs before measurementUpdate, the initial value set in KalmanFilter is
never used.

Not only that, but it isn't scaled by unitConversion, when this->state is
scaled and is otherwise just a copy of initialState like xBar is.