Return function calls to use "To" preposition
These functions were erroneously renamed under the expectation
that they would return views on memory (Eigen:Map). This is
not the case and they are copies and should remain as copies.
They should remain as copies because the size of the returned
vector/matrix values are small in memory terms.